cleaned up code

+++ controllers/userController.js
@@ -47,14 +47,19 @@ module.exports = {
     },
     async deleteUser(req, res) {
         try {
-            const user = await User.findOneAndDelete(
-            { users: req.params.userId },
-            { $pull: { users: req.params.userId } },
-            { new: true }
-            );
+            const user = await User.findOneAndDelete({ _id: req.params.userId });
             if (!user) {
-                return res.status(404).json({ message: 'No user found at this ID' });
+                return res.status(404).json({ message: 'No such User exists' });
             }
+
+            // const user = await User.findOneAndDelete(
+            // { users: req.params.userId },
+            // { $pull: { users: req.params.userId } },
+            // { new: true }
+            // );
+            // if (!user) {
+            //     return res.status(404).json({ message: 'No user found at this ID' });
+            // }
             res.json({ message: 'User successfully deleted' });
         } catch (err) {
             res.status(500).json(err);
@@ -64,7 +69,7 @@ module.exports = {
         try {
             const user = await User.findOneAndUpdate(
                 { _id: req.params.userId },
-                { $addToSet: { friends: req.body.friendId || req.params.friendId } },
+                { $addToSet: { friends: req.body.friendId } },
                 { new: true }
             );
             if (!user) {
@@ -77,10 +82,12 @@ module.exports = {
     },
     async removeFriend(req, res) {
         try {
+            console.log(req);
             const user = await User.findOneAndUpdate(
                 { _id: req.params.userId },
-                { $pull: { friends: params.friendId } },
-                { new: true }
+                // { $pull: { friends: { friendId: req.params.friendId } } },
+                { $pull: { friends: req.params.friendId } },
+                { runValidators: true, new: true }
             );
             if (!user) {
                 return res.status(404).json({ message: 'No user found at this ID' });
